requirements.txt
anyio==3.5.0
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blis==0.7.7
bpemb==0.3.3
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
conllu==4.4.1
cycler==0.11.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.1
Deprecated==1.2.13
en-core-web-sm==3.2.0
entrypoints==0.4
executing==0.8.3
fastapi==0.75.0
filelock==3.6.0
flair==0.11
fonttools==4.31.2
ftfy==6.1.1
future==0.18.2
gdown==3.12.2
gensim==4.1.2
h11==0.13.0
httptools==0.4.0
huggingface-hub==0.4.0
hyperopt==0.2.7
idna==3.3
importlib-metadata==3.10.1
inflect==5.4.0
iniconfig==1.1.1
ipykernel==6.9.2
ipython==8.1.1
Janome==0.4.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.0
konoha==4.6.5
langcodes==3.3.0
langdetect==1.0.9
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==8.8.0
mpld3==0.3
murmurhash==1.0.6
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.8
nltk==3.7
numpy==1.22.3
overrides==3.1.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.4
plotly==5.6.0
pluggy==1.0.0
pptree==3.1
preshed==3.0.6
prompt-toolkit==3.0.27
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.5
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
segtok==1.5.11
sentence-transformers==2.2.0
sentencepiece==0.1.95
setuptools==60.10.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
sqlitedict==2.0.0
srsly==2.4.2
stack-data==0.2.0
starlette==0.17.1
tabulate==0.8.9
tenacity==8.0.1
thinc==8.0.15
thinc-apple-ops==0.0.5
threadpoolctl==3.1.0
tokenizers==0.11.6
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
transformers==4.17.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
wasabi==0.9.0
watchgod==0.8.1
wcwidth==0.2.5
websockets==10.2
wheel==0.37.1
Wikipedia-API==0.5.4
word2number==1.1
wrapt==1.14.0
zipp==3.7.0
